:shipit: Render fiddles via `hyperfiddle.api/render-fiddle` multimethod
Also drop `eval-domain-code+` (and so `:project/code`). Using `eval` to render
projects and fiddles was required when we used to store code as strings in
datomic. This is not needed anymore now that fiddles are defined in source
control. This removes some tech debt and improve development time
performances (faster code reloading).