Fix failing test because of datomic internal difference
I'm using datomic free, the CI pipeline is using datomic pro. For some reason,
and internal difference resulted in query result to be returned in a different
order. Tests were failing because `{:foo [:a :b]} ≠ {:foo [:b :a]}`.

Since it's not trivial to compare deeply nested values while ignoring ordering,
I introduced the `'in-any-order` `clojure.test` assertion.
```
(is (in-any-order {:foo [:a :b]} {:foo [:b :a]})) ;;=> true
```